+++ common/src/main/java/me/pandamods/fallingtrees/api/TreeDetectionAlgorithm.java
@@ -1,0 +1,10 @@
+package me.pandamods.fallingtrees.api;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+
+import java.util.Set;
+
+public abstract class TreeDetectionAlgorithm {
+	public abstract Set<BlockPos> gatherTreeBlocks(Level level, BlockPos startPos);
+}