+++ CHANGELOG.md
@@ -12,10 +12,10 @@
 * Changed `maxLeavesRadius` default value 10 -> 7
 * Removed `shouldFallOnMaxLogAmount` as it became redundant with the new algorithm changes to Generic Tree
 * Renamed `onlyFallWithRequiredTool` to `requireTool`
+* Removed `mushroomTree` filters, this is now handled internally
 
 ### Algorithm Update
-#### Chorus Plant
-* Minor improvements
+* Highly improved the Generic Tree's tree detection algorithm by using the Breadth-first search (BFS) method.
 * Chorus Plants will now only fall from the mined block and above, leaving the blocks under the mined block.
 
 ## Developer Notes
@@ -26,3 +26,4 @@
 * Renamed `standardTree.class` to `genericTree.class`
 * Renamed `standardTreeConfig.class` to `genericTreeConfig.class`
 * Renamed `Tree.class` to `TreeType.class`
+* Split `MushroomTree.class` up into `RedMushroomTree.class` and `BrownMushroomTree.class`